requirements.txt
certifi==2020.6.20
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
idna==2.8
numpy==1.17.3
pandas==0.25.3
pycparser==2.19
pyOpenSSL==19.0.0
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
selenium==3.141.0
six==1.13.0
urllib3==1.25.7